requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.cached-property==1.0.1
botocore==1.26.5
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
frozenlist==1.3.0
gql==3.3.0
graphql-core==3.2.1
h11==0.13.0
idna==3.3
jmespath==1.0.0
multidict==6.0.2
Pygments==2.12.0
python-dateutil==2.8.2
python-multipart==0.0.5
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sniffio==1.2.0
starlette==0.20.0
strawberry-graphql==0.113.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
websockets==10.3
yarl==1.7.2
